Add info and use scale-ci dedicated agents/slaves
This commit sets the default Node label in the pipeline to scale-ci. We have
multiple slaves which are always online instead of the on demand slaves which
take time to get ready before scheduling the jobs. There are multiple slaves
and executors, so we should be able to run jobs in parallel.