Use kubeconfig instead of oc login
This commit enables us to run uperf tests using the kubeconfig copied
from the orchestration host instead of logging in as kubeadmin user.
This way we need not parse the cluster api url, kubeadmin password
during the install job and pass them to the uperf job in order to
login to fetch kubeconfig on the slave after building the ocp cluster.