Enable watcher tracking for all the pipelines
This commit adds the jjb templates for all the pipelines part of the
PerfScale CI for watcher to be able to track/make changes. This way
everything is automated. Any change to the default workloads enabled
and the properties files can be made through github similar to how it's
done for the workloads.